+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Versioning for this project works like this: [Versioning](/docs/details/05-Versi
 ## v0.2.1 - ...
 
 ### Added
+- Added automatic generation of function import declarations when compiling static and dynamic libraries.
 - Automatic configuration of MSVC environment when running compiler.
 - Experimental ARM code generation
 - Added *#function* which is evaluated to the name of the current function, similar to *#line*.